+++ src/afx2lg/lg/lg_parser.h
@@ -18,7 +18,7 @@ using std::tr1::shared_ptr;
 class LgParserCallback {
  public:
   virtual const axefx::PresetMap& GetPresetMap() = 0;
-  virtual void WriteLine(const char* line, int length) = 0;
+  virtual void WriteLine(const char* line, size_t length) = 0;
 };
 
 class LgParser {
